% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF-fn.R
\name{deviance.backf}
\alias{deviance.backf}
\title{Deviance for objects of class \code{backf}}
\usage{
\method{deviance}{backf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\value{
A real number.
}
\description{
This function returns the deviance of the fitted additive model using one of the three
classical or robust marginal integration estimators, as computed with \code{\link{backf.cl}} or
\code{\link{backf.rob}}.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
