% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF-fn.R
\name{fitted.values.backf}
\alias{fitted.values.backf}
\title{Fitted values for objects of class \code{backf}}
\usage{
fitted.values.backf(object, ...)
}
\arguments{
\item{object}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\value{
A vector of fitted values.
}
\description{
This function returns the fitted values given the covariates of the original sample under an additive model using a classical or robust marginal integration procedure estimator computed with \code{backf.cl} or \code{backf.rob}.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
