% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF-fn.R
\name{plot.backf}
\alias{plot.backf}
\title{Diagnostic plots for objects of class \code{backf}}
\usage{
\method{plot}{backf}(object, which = 1:np, ask = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}

\item{which}{vector of indices of explanatory variables for which partial residuals plots will
be generaetd. Defaults to all available explanatory variables.}

\item{ask}{logical value. If \code{TRUE}, the graphical device will prompt for confirmation before
going to the next page/screen of output.}

\item{...}{additional other arguments. Currently ignored.}
}
\description{
Plot method for objects of class \code{backf}.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
