% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF-fn.R
\name{print.backf}
\alias{print.backf}
\title{Print a Marginal Integration procedure}
\usage{
\method{print}{backf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{backf}, a result of a call to \code{\link{backf.cl}} or \code{\link{backf.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\value{
A real number.
}
\description{
The default print method for a \code{backf} object.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
