\name{RBF-package}
\alias{RBF-package}
\alias{RBF}
\docType{package}
\title{
A robust backfitting algorithm for additive models.
}
\description{
A robust backfitting algorithm for additive models.
}

\author{
Matias Salibian-Barrera, Alejandra Martinez

Maintainer: Matias Salibian-Barrera <matias@stat.ubc.ca>
}
\references{
Boente G, Martinez A, Salibian-Barrera M. Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics, 2017; 29:744-767. https://doi.org/10.1080/10485252.2017.1369077
}
\keyword{ Robust Backfitting }
