% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf.R
\name{rbf}
\alias{rbf}
\alias{predict.rbf}
\title{Radial basis function model}
\usage{
rbf(X, Y, augmented = TRUE, kernel = "cubic")

\method{predict}{rbf}(object, xp, ...)
}
\arguments{
\item{X}{Matrix witn inputs}

\item{Y}{Matrix with outputs}

\item{augmented}{Should the augmented RBF model be used?}

\item{kernel}{Which kernel should be used}

\item{object}{RBF model}

\item{xp}{Matrix of points to predict at}

\item{...}{Further arguments passed to predict}
}
\value{
RBF model

Predictions at xp
}
\description{
Radial basis function model

Predict with RBF
}
\examples{
d <- 8
n <- 300
x <- matrix(runif(d*n), ncol=d)
f <- function(x) {cos(x[1]^1.9)}
y <- as.matrix(apply(x, 1, f))
arb1 <- rbf(X=x, Y = y)
np <- 300
xp <- matrix(runif(d*np), ncol=d)
p1 <- predict(arb1, xp)
}
