% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF-fn.R
\name{summary.backf}
\alias{summary.backf}
\alias{summary.backf.cl}
\alias{summary.backf.rob}
\title{Summary for additive models fits using backfitting}
\usage{
\method{summary}{backf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{backf}, a result of a call to
\code{\link{backf.cl}} or \code{\link{backf.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\description{
Summary method for class \code{backf}.
}
\details{
This function returns the estimation of the intercept and also the
five-number summary and the mean of the residuals for both classical and
robust estimators. For the classical estimator, it also returns the R-squared.
For the robust estimator it returns a robust version of the R-squared and 
the estimate of the residual standard error.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
