\name{RBF-package}
\alias{RBF-package}
\alias{RBF}
\docType{package}
\title{
A robust backfitting algorithm for additive models.
}
\description{
A robust backfitting algorithm for additive models.
}
\details{
\tabular{ll}{
Package: \tab RBF\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-19\cr
License: \tab GPL 3.0\cr
}
}
\author{
Matias Salibian-Barrera, Alejandra Martinez

Maintainer: Matias Salibian-Barrera <matias@stat.ubc.ca>
}
\references{
Boente G, Martinez A, Salibian-Barrera M. Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics, 2017; 29:744-767. https://doi.org/10.1080/10485252.2017.1369077
}
\keyword{ Robust Backfitting }
